Rename old_model.py to model_old.py
